requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
# Agent framework
langgraph==0.2.35
# Optional embeddings retrieval (guarded by MODE=online)
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
# GCP clients (optional for Firestore/GCS; not required for local run)
google-cloud-firestore==2.16.0
google-cloud-storage==2.18.2
